+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Revert elastic search. Refs UIIN-1822.
 * Use correct `css-loader` syntax. Refs UIIN-1826.
 * Fix suppressed from discovery filter. Fixes UIIN-1832.
+* Create a MARC Holdings Record. Refs UIIN-1828.
 
 ## [8.0.0](https://github.com/folio-org/ui-inventory/tree/v8.0.0) (2021-10-05)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v7.1.4...v8.0.0)